+++ test/unit/mysql_db_admin/data_out.py
@@ -110,6 +110,9 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
+        test_outfile_mode_expand2
+        test_outfile_mode_expand
+        test_outfile_expand
         test_outfile_mode2
         test_outfile_mode
         test_outfile
@@ -160,6 +163,70 @@ class UnitTest(unittest.TestCase):
         self.results2 = (
             False, "Error: Is not a dictionary: %s" % (self.data2))
         self.results3 = (False, "Error Message")
+        self.outfile = "path/to/open"
+
+    @mock.patch("mysql_db_admin.pprint.pprint", mock.Mock(return_value=True))
+    @mock.patch("builtins.open", new_callable=mock.mock_open, read_data="data")
+    def test_outfile_mode_expand2(self, mock_file):
+
+        """Function:  test_outfile_mode_expand2
+
+        Description:  Test with outfile and mode and expand options.
+
+        Arguments:
+
+        """
+
+        assert open(                            # pylint:disable=R1732,W1514
+            self.outfile).read() == "data"
+        mock_file.assert_called_with(self.outfile)
+
+        self.assertEqual(
+            mysql_db_admin.data_out(
+                self.data, suppress=self.suppress, outfile=self.outfile,
+                mode=self.mode2, expand=True), self.results)
+
+    @mock.patch("mysql_db_admin.pprint.pprint", mock.Mock(return_value=True))
+    @mock.patch("builtins.open", new_callable=mock.mock_open, read_data="data")
+    def test_outfile_mode_expand(self, mock_file):
+
+        """Function:  test_outfile_mode_expand
+
+        Description:  Test with outfile and mode and expand options.
+
+        Arguments:
+
+        """
+
+        assert open(                            # pylint:disable=R1732,W1514
+            self.outfile).read() == "data"
+        mock_file.assert_called_with(self.outfile)
+
+        self.assertEqual(
+            mysql_db_admin.data_out(
+                self.data, suppress=self.suppress, outfile=self.outfile,
+                mode=self.mode, expand=True), self.results)
+
+    @mock.patch("mysql_db_admin.pprint.pprint", mock.Mock(return_value=True))
+    @mock.patch("builtins.open", new_callable=mock.mock_open, read_data="data")
+    def test_outfile_expand(self, mock_file):
+
+        """Function:  test_outfile_expand
+
+        Description:  Test with outfile option with expand.
+
+        Arguments:
+
+        """
+
+        assert open(                            # pylint:disable=R1732,W1514
+            self.outfile).read() == "data"
+        mock_file.assert_called_with(self.outfile)
+
+        self.assertEqual(
+            mysql_db_admin.data_out(
+                self.data, suppress=self.suppress, outfile=self.outfile
+                , expand=True), self.results)
 
     @mock.patch("mysql_db_admin.gen_libs.write_file",
                 mock.Mock(return_value=True))